requirements.txt
feedparser==6.0.11
requests==2.31.0
groq==0.28.0
google-genai==1.10.0
google-cloud-texttospeech
google-cloud-videointelligence
edge-tts
gTTS==2.5.3
google-api-python-client==2.136.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
faster-whisper==1.0.3
yfinance==0.2.54
pillow
pandas
fredapi
matplotlib
openai
